task layout: display editions task
