+++ app/modules/tasks/tasks.ts
@@ -8,6 +8,7 @@ export default {
         'tasks(/)(humans)(/)': 'showHumansTask',
         'tasks(/)(publishers)(/)': 'showPublisherTask',
         'tasks(/)(works)(/)': 'showWorksTask',
+        'tasks(/)(editions)(/)': 'showEditionsTask',
         'tasks(/)(:id)(/)': 'showTask',
       },
     })
@@ -21,6 +22,7 @@ const controller = {
   showHumansTask (task) { controller.showTask(task, 'human') },
   showPublisherTask (task) { controller.showTask(task, 'publisher') },
   showWorksTask (task) { controller.showTask(task, 'work') },
+  showEditionsTask (task) { controller.showTask(task, 'edition') },
   showTask (task, type) {
     if (app.request('require:dataadmin:access', 'tasks')) {
       return showLayout({ task, entitiesType: type })
